Add Podman API and restructure pod.py component
